修改编译不过

+++ server/gopath/src/gitee.com/jntse/minehero/server/gateserver/useritem.go
@@ -32,7 +32,7 @@ func (this *GateUser) AddItem(item uint32, num uint32, reason string) {
 		this.bag.AddItem(item, num, reason)
 	}
 	//CountMgr().AddGet(item, num)
-	RCounter().IncrByDate("item_add", uint32(item), num)
+	RCounter().IncrByDate("item_add", int32(item), int32(num))
 }
 
 // 扣除道具
@@ -59,7 +59,7 @@ func (this *GateUser) RemoveMoney(gold uint32, reason string, syn bool) bool {
 			this.SendMsg(send)
 		}
 		log.Info("玩家[%d] 扣除金币[%d] 剩余[%d] 原因[%s]", this.Id(), gold, this.GetMoney(), reason)
-		RCounter().IncrByDate("item_remove", uint32(msg.ItemId_Gold), gold)
+		RCounter().IncrByDate("item_remove", int32(msg.ItemId_Gold), int32(gold))
 		return true
 	}
 	log.Info("玩家[%d] 扣除金币失败[%d] 原因[%s]", this.Id(), gold, reason)
@@ -72,7 +72,7 @@ func (this *GateUser) AddYuanbao(yuanbao uint32, reason string) {
 	this.yuanbao = this.GetYuanbao() + yuanbao
 	send := &msg.GW2C_UpdateYuanbao{Num:pb.Uint32(this.GetYuanbao())}
 	this.SendMsg(send)
-	RCounter().IncrByDate("item_add", uint32(msg.ItemId_YuanBao), yuanbao)
+	RCounter().IncrByDate("item_add", int32(msg.ItemId_YuanBao), int32(yuanbao))
 	log.Info("玩家[%d] 添加元宝[%d] 库存[%d] 原因[%s]", this.Id(), yuanbao, this.GetYuanbao(), reason)
 }
 func (this *GateUser) RemoveYuanbao(yuanbao uint32, reason string) bool {
@@ -81,7 +81,7 @@ func (this *GateUser) RemoveYuanbao(yuanbao uint32, reason string) bool {
 		send := &msg.GW2C_UpdateYuanbao{Num:pb.Uint32(this.GetYuanbao())}
 		this.SendMsg(send)
 		log.Info("玩家[%d] 扣除元宝[%d] 库存[%d] 原因[%s]", this.Id(), yuanbao, this.GetYuanbao(), reason)
-		RCounter().IncrByDate("item_remove", uint32(msg.ItemId_YuanBao), yuanbao)
+		RCounter().IncrByDate("item_remove", int32(msg.ItemId_YuanBao), int32(yuanbao))
 		return true
 	}
 	log.Info("玩家[%d] 扣除元宝[%d]失败 库存[%d] 原因[%s]", this.Id(), yuanbao, this.GetYuanbao(), reason)
@@ -104,7 +104,7 @@ func (this *GateUser) RemoveCoupon(num uint32, reason string) bool {
 		this.SendMsg(send)
 		log.Info("玩家[%d] 添加金卷[%d] 库存[%d] 原因[%s]", this.Id(), num, this.GetCoupon(), reason)
 		//CountMgr().AddRemove(uint32(msg.ItemId_Coupon), num)
-		RCounter().IncrByDate("item_remove", uint32(msg.ItemId_Coupon), num)
+		RCounter().IncrByDate("item_remove", int32(msg.ItemId_Coupon), int32(num))
 		return true
 	}
 	log.Info("玩家[%d] 添加金卷[%d]失败 库存[%d] 原因[%s]", this.Id(), num, this.GetCoupon(), reason)
